etl/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.37
psycopg[binary]==3.2.3
pandas==2.2.3
numpy==2.1.3
networkx==3.4.2
